fix(thread) : fixed extraReducers delete thread

+++ src/components/table/ThreadTable.jsx
@@ -1,29 +1,31 @@
-import React, { useState } from "react";
+import React from "react";
 import { Button, Popover, Table } from "antd";
 import { InfoCircleOutlined, DeleteOutlined } from "@ant-design/icons";
 import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteThread, getThread } from "../../store/thread/ThreadSlicer";
+import moment from "moment";
+import "moment/locale/id";
+import Swal from "sweetalert2";
 
 const ThreadTable = (props) => {
-  const response = useSelector((state) => state?.thread.data.threads);
-  const [infoKeyId, setInfoKeyId] = useState("");
-  const isEditing = (record) => record.key === infoKeyId;
   const navigate = useNavigate();
   const dispacth = useDispatch();
-  console.log("response ", response);
-
-  const getInfo = (record) => {
-    setInfoKeyId(record.key);
-  };
-  //
-  const cancelDetail = (record) => {
-    setInfoKeyId("");
-  };
 
   const onDeleteHandler = (id) => {
-    dispacth(deleteThread(id));
-    cancelDetail();
+    Swal.fire({
+      title: "Are you sure?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire("Deleted!", "Topic has been deleted.", "success");
+        dispacth(deleteThread(id));
+      }
+    });
   };
   //
   const columns = [
@@ -47,12 +49,16 @@ const ThreadTable = (props) => {
     {
       title: "Topic",
       dataIndex: "topic",
-      render: (data) => data.topic,
+      render: (val) => val.topic,
       width: "10%",
     },
     {
       title: "Date",
       dataIndex: "createdAt",
+      render: (val) => {
+        moment.locale("id");
+        return moment(val).format("ll");
+      },
       width: "10%",
     },
     {
@@ -61,53 +67,35 @@ const ThreadTable = (props) => {
       width: "10%",
       editable: true,
       render: (_, record) => {
-        const infoDetail = isEditing(record);
         return (
           <>
-            {infoDetail ? (
-              <Popover
-                defaultOpen={infoDetail}
-                content={
-                  <div style={{ display: "grid" }}>
-                    <Button
-                      type="text"
-                      style={{ marginBottom: "10px", background: "#D1E6E0" }}
-                      onClick={() => {
-                        dispacth(getThread(record._id));
-                        navigate("/dashboard/thread/" + record._id);
-                      }}
-                    >
-                      <InfoCircleOutlined />
-                      Details
-                    </Button>
-                    <Button
-                      type="text"
-                      onClick={() => onDeleteHandler(record._id)}
-                    >
-                      <DeleteOutlined />
-                      Delete
-                    </Button>
-                  </div>
-                }
-                destroyTooltipOnHide
-              >
-                <Button
-                  disabled={infoKeyId === ""}
-                  onClick={() => cancelDetail()}
-                  type={"text"}
-                >
-                  &#x2022; &#x2022; &#x2022;
-                </Button>
-              </Popover>
-            ) : (
-              <Button
-                disabled={infoKeyId !== ""}
-                onClick={() => getInfo(record)}
-                type={"text"}
-              >
-                &#x2022; &#x2022; &#x2022;
-              </Button>
-            )}
+            <Popover
+              content={
+                <div style={{ display: "grid" }}>
+                  <Button
+                    type="text"
+                    style={{ marginBottom: "10px", background: "#D1E6E0" }}
+                    onClick={() => {
+                      dispacth(getThread(record._id));
+                      navigate("/dashboard/thread/" + record._id);
+                    }}
+                  >
+                    <InfoCircleOutlined />
+                    Details
+                  </Button>
+                  <Button
+                    type="text"
+                    onClick={() => onDeleteHandler(record._id)}
+                  >
+                    <DeleteOutlined />
+                    Delete
+                  </Button>
+                </div>
+              }
+              destroyTooltipOnHide
+            >
+              <Button type={"text"}>&#x2022; &#x2022; &#x2022;</Button>
+            </Popover>
           </>
         );
       },
@@ -130,13 +118,7 @@ const ThreadTable = (props) => {
   return (
     <Table
       bordered
-      dataSource={
-        // props.response?.data?
-        // value
-        // :
-        // null
-        response
-      }
+      dataSource={props?.response}
       columns={mergedColumns}
       rowClassName="editable-row"
       rowKey={(val) => val._id}
